color palette and improve alternate screen

+++ src/ui/macos-ui/macos-ui/Terminal/Models/TerminalCell.swift
@@ -5,7 +5,7 @@ struct TerminalCell: Identifiable {
     let codepoint: UInt32
     let isEmpty: Bool
     let isBold: Bool
-    let color: UInt8
+    let color: Int32
 
     var character: String {
         guard !isEmpty, let scalar = UnicodeScalar(codepoint) else {
